Added license

+++ Unit 3: Boolean Expressions and If Statements/3.2.8: Cooking/MicrowaveCooking.java
@@ -1,19 +1,36 @@
+/* MIT License
+ *
+ * Copyright (c) 2021 Brighton Sikarskie
+ *  
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ * 
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ * 
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
 public class MicrowaveCooking
 {
     public static void main(String[] args)
     {
-       // Generate a random number of seconds
-       // between 0 and 60
-
-       int seconds = (int) ((double)Math.random() * 60);
-       // Print the number of seconds
+       int seconds = (int) (Math.random() * 60);
        System.out.println("Microwaving for " + seconds + " seconds");
-       // Use two if statements to print
        if (seconds <= 20) {
            System.out.println("Perfect cooking time!");
        } else {
            System.out.println("Your roll will catch fire!");
        }
-       // whether the roll is fine or will catch fire
     }
 }
